refactor: flatten search endpoints
