+++ api/snapshots/search.ts
@@ -1,7 +1,7 @@
 // Moved to prevent route collision with [id].ts in Next.js
-const getAirtableContext = require("../../../lib/airtableBase");
-const { createSearchHandler } = require("../../../lib/airtableSearch");
-const { getFieldMap } = require("../../../lib/resolveFieldMap");
+const getAirtableContext = require("../../lib/airtableBase");
+const { createSearchHandler } = require("../../lib/airtableSearch");
+const { getFieldMap } = require("../../lib/resolveFieldMap");
 
 const apiSnapshotsSearchHandler = async (req: any, res: any) => {
   const { TABLES } = getAirtableContext();
